litex: Fix SoC hanging on UART stub
In LiteX a 'stub' uart is used to create the CPU interface for a UART
but detach it from any physical interface. It's created with a stream
interface with the transmit path's ready sigal fixed HIGH. However
there is a bug, it's the rx path's ready signal that is set. This
causes the UART to stall the SoC, because it is blocked from sending
data.

Fixes #13